requirements.txt
streamlit
requests
google-generativeai==0.8.5
protobuf==4.25.3
google-auth==2.29.0
streamlit-lottie
